--- a/data.sql
+++ b/data.sql
@@ -167,9 +167,4 @@ INSERT INTO visits (animal_id, vet_id, date_of_visit) SELECT * FROM (SELECT id F
 -- This will add 2.500.000 owners with full_name = 'Owner <X>' and email = 'owner_<X>@email.com' (~2min approx.)
 insert into owners (full_name, email) select 'Owner ' || generate_series(1,2500000), 'owner_' || generate_series(1,2500000) || '@mail.com';
 
--- Creating an INDEX
-CREATE INDEX visits(animal_vet_id_idx) ON visits(animal_id, vet_id);
-CREATE INDEX email_idx ON owners(email);
-
-