--- a/queries.sql
+++ b/queries.sql
@@ -125,6 +125,11 @@ WHERE ve.name = 'Maisy Smith'
 ORDER BY v.visit_date ASC
 LIMIT 1;
 
+-- ================================================================================================================
+-- Performance
+-- Add an email column to your owners table
+ALTER TABLE owners ADD COLUMN email VARCHAR(120);
+