## Database performance audit.
- Improved execution time for each query.
- Applied indexing

|No.| Query :recycle:   | Before Indexing(ms) :slightly_frowning_face:  | After Indexing(ms) :slightly_smiling_face:  |
|---| ------------- | ------------- | ------------- |
|1.| SELECT COUNT(*) FROM visits where animal_id = 4; | 174.078 |57.323 |
|2.|SELECT * FROM visits where vet_id = 2; | 321.501  |314.616  |
|3.| SELECT * FROM owners where email = 'owner_18327@mail.com' | 155.555  | 21.833  |

![Screenshot from 2023-02-07 12-24-19](https://user-images.githubusercontent.com/61404403/217204426-6d19ea4e-4b31-45bb-adcc-2d811395a765.png)
![Screenshot from 2023-02-07 12-23-34](https://user-images.githubusercontent.com/61404403/217204442-afc7fe5d-a4b5-42ed-b6dc-9b0aa0593727.png)
![Screenshot from 2023-02-07 12-22-58](https://user-images.githubusercontent.com/61404403/217204446-9c8ecce4-79c3-47fd-b973-01702453d975.png)
![Screenshot from 2023-02-07 11-25-35](https://user-images.githubusercontent.com/61404403/217204454-669efcd3-c57c-4830-9ab7-e0e5c7a452d9.png)
![Screenshot from 2023-02-07 11-25-11](https://user-images.githubusercontent.com/61404403/217204457-af81f772-658e-40b7-a0f5-7d4a7033136f.png)
![Screenshot from 2023-02-07 02-50-54](https://user-images.githubusercontent.com/61404403/217204470-e7937557-bc23-4796-b923-f0e362e686ed.png)
![Screenshot from 2023-02-07 02-52-14](https://user-images.githubusercontent.com/61404403/217204478-2a82fa59-e7dd-4096-80db-9e26b6c12bdc.png)
![Screenshot from 2023-02-07 02-51-35](https://user-images.githubusercontent.com/61404403/217204488-8f934a12-fe8b-4689-ab98-621bd3004e6a.png)
![Screenshot from 2023-02-07 02-46-24](https://user-images.githubusercontent.com/61404403/217204491-761fa226-f390-42e3-8d9f-34daad2851c5.png)